[FE] REFACTOR: member list 컴포넌트 분리
