+++ src/frontend/src/components/Sidebar/MemberList/MemberListFooter/index.tsx
@@ -1,0 +1,47 @@
+import { useState } from 'react';
+import { CommonInput } from '@/components/common/Input';
+import { MemberFooter, VoiceChatFooter, ActionButton, JoinButton } from '../index.css';
+
+import { SidebarType } from '@/types/enums/SidebarType';
+
+import AddUserIcon from '@/assets/img/AddUser.svg';
+import MicrophoneOn from '@/assets/img/MicrophoneOn.svg';
+import HeadphoneOn from '@/assets/img/HeadphoneOn.svg';
+import MicrophoneOffRed from '@/assets/img/MicrophoneOffRed.svg';
+import HeadphoneOffRed from '@/assets/img/HeadphoneOffRed.svg';
+
+interface IMemberListFooter {
+  sidebarType: SidebarType;
+}
+
+export const MemberListFooter = (props: IMemberListFooter) => {
+  const [micOn, setMicOn] = useState(true);
+  const [soundOn, setSoundOn] = useState(true);
+
+  const handleMicrophone = () => setMicOn(!micOn);
+  const handleSound = () => setSoundOn(!soundOn);
+
+  if (props.sidebarType === SidebarType.MEMBER) {
+    return (
+      <MemberFooter>
+        <img src={AddUserIcon} alt="Add User" />
+        <CommonInput placeholder="룸 유저 검색" design={1} />
+      </MemberFooter>
+    );
+  }
+
+  if (props.sidebarType === SidebarType.VOICECHAT) {
+    return (
+      <VoiceChatFooter>
+        <ActionButton onClick={handleMicrophone}>
+          <img src={micOn ? MicrophoneOn : MicrophoneOffRed} />
+        </ActionButton>
+        <ActionButton onClick={handleSound}>
+          <img src={soundOn ? HeadphoneOn : HeadphoneOffRed} />
+        </ActionButton>
+        <JoinButton>입장</JoinButton>
+      </VoiceChatFooter>
+    );
+  }
+  return null;
+};
